Split livesession source files to pair with test files
