+++ internal/livesession/livesession.go
@@ -757,223 +757,6 @@ func (s *Service) StartQuiz(ctx context.Context, joinCode string, hostPlayerID, 
 	return nil
 }
 
-// StartHosting opens or reuses the host's live room for a quiz, enforcing one
-// room per host (#851). It backs the quiz-view "Host live" control:
-//   - No active room -> open a new lobby armed with the quiz (host starts it
-//     when players are in), same as the prior "Play live".
-//   - Active empty staging lobby OR between-games intermission -> arm the quiz
-//     in it but stay in the lobby (reusing ArmQuiz), so the host gathers players
-//     and presses Start, the same as the no-active-room case (#863, #875). No
-//     second room is spawned, and the post-game pick never auto-starts.
-//   - Active room with a game in flight -> leave it untouched and return it, so
-//     a stray pick never disrupts a running game (the end-and-restart confirm
-//     is deferred to #853).
-//
-// It returns the room the host should be redirected to; only its JoinCode is
-// authoritative (the redirect target). In the reuse branch the returned snapshot
-// predates the arm, so its Phase/QuizID/StartedAt may lag the now-armed room -
-// callers must re-read if they need post-arm state. [quiz.ErrQuizNotFound] and
-// [ErrNotLiveQuiz] propagate so the handler can bounce an unhostable quiz to the
-// quiz list.
-func (s *Service) StartHosting(ctx context.Context, quizID, hostPlayerID int64) (*Session, error) {
-	active, err := s.store.GetActiveSessionForHost(ctx, hostPlayerID)
-	if err != nil {
-		return nil, fmt.Errorf(errGetActiveSessionFmt, err)
-	}
-
-	if active == nil {
-		var sess *Session
-		if sess, err = s.CreateSession(ctx, &quizID, hostPlayerID); err != nil {
-			return nil, err
-		}
-		s.logger.InfoContext(ctx, "host started hosting: opened new room",
-			slog.String(logJoinCodeKey, sess.JoinCode),
-			slog.Int64(logHostKey, hostPlayerID),
-			slog.Int64(logQuizKey, quizID))
-
-		return sess, nil
-	}
-
-	if !canArmQuiz(active) {
-		// A game is in flight: leave the running room untouched and return it so
-		// a stray pick never disrupts it. The end-and-restart confirm is #853.
-		s.logger.InfoContext(ctx, "host started hosting: reused running room",
-			slog.String(logJoinCodeKey, active.JoinCode),
-			slog.Int64(logHostKey, hostPlayerID),
-			slog.Int64(logQuizKey, quizID))
-
-		return active, nil
-	}
-
-	// An empty staging lobby and the between-games intermission both take a new
-	// quiz without spawning a second room, and both arm it and stay in the lobby
-	// so the host gathers players and presses Start (#875), matching the
-	// no-active-session case which also lands on an armed lobby. canArmQuiz
-	// guarantees active is one of these two; ArmQuiz handles both - RearmSession
-	// re-arms an intermission room back to an armed lobby with started_at cleared.
-	err = s.ArmQuiz(ctx, active.JoinCode, hostPlayerID, quizID)
-	switch {
-	case err == nil, errors.Is(err, ErrGameInFlight):
-		// ErrGameInFlight means the room raced into flight between the read above
-		// and the arm; treat it the same as "in flight, do nothing" and return
-		// the room so the host lands on it.
-		s.logger.InfoContext(ctx, "host started hosting: reused active room",
-			slog.String(logJoinCodeKey, active.JoinCode),
-			slog.Int64(logHostKey, hostPlayerID),
-			slog.Int64(logQuizKey, quizID),
-			slog.String(logPhaseKey, string(active.Phase)))
-
-		return active, nil
-	default:
-		return nil, err
-	}
-}
-
-// HostHasRunningGame reports whether the host has an active session with a game
-// in flight (#853): one that has left the lobby/intermission staging and so
-// cannot take a new quiz without ending it. The quiz view uses this to gate the
-// "Host live" confirm-and-restart prompt. An empty staging lobby, an
-// armed-but-not-started lobby, and the between-games intermission all return
-// false (a pick just arms them, no confirm needed).
-func (s *Service) HostHasRunningGame(ctx context.Context, hostPlayerID int64) (bool, error) {
-	active, err := s.store.GetActiveSessionForHost(ctx, hostPlayerID)
-	if err != nil {
-		return false, fmt.Errorf(errGetActiveSessionFmt, err)
-	}
-
-	return active != nil && !canArmQuiz(active), nil
-}
-
-// RestartHosting ends the host's active session (if any) and opens a new room
-// hosting quizID (#853) - the deliberate "switch the live quiz" path the host
-// confirms when a game is already running. The target quiz is validated BEFORE
-// the running session is ended, so an unhostable pick ([quiz.ErrQuizNotFound] /
-// [ErrNotLiveQuiz]) bounces with nothing torn down. A rarer failure of the
-// create after the end (e.g. join-code exhaustion) surfaces as an error; the old
-// session is already gone, so the host's retry simply opens the new room.
-func (s *Service) RestartHosting(ctx context.Context, quizID, hostPlayerID int64) (*Session, error) {
-	// Validate the target quiz up front; CreateSession re-checks it, but doing it
-	// here first guarantees we never end the running game for an unhostable pick.
-	qz, err := s.quizzes.GetQuiz(ctx, quizID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get quiz for restart: %w", err)
-	}
-	if qz.Mode != quiz.ModeLive {
-		return nil, ErrNotLiveQuiz
-	}
-
-	active, err := s.store.GetActiveSessionForHost(ctx, hostPlayerID)
-	if err != nil {
-		return nil, fmt.Errorf(errGetActiveSessionFmt, err)
-	}
-	if active != nil {
-		if err = s.EndSession(ctx, active.JoinCode, hostPlayerID); err != nil {
-			return nil, fmt.Errorf("failed to end running session: %w", err)
-		}
-	}
-
-	// No active session now (just ended, or never any): CreateSession opens a new
-	// armed lobby hosting the picked quiz.
-	var sess *Session
-	if sess, err = s.CreateSession(ctx, &quizID, hostPlayerID); err != nil {
-		return nil, err
-	}
-	s.logger.InfoContext(ctx, "host restarted hosting: opened new room",
-		slog.String(logJoinCodeKey, sess.JoinCode),
-		slog.Int64(logHostKey, hostPlayerID),
-		slog.Int64(logQuizKey, quizID))
-
-	return sess, nil
-}
-
-// ArmQuiz arms a quiz to play in a room but leaves it in the lobby, not started
-// (#863): the host then begins it with the existing "Start now" control once
-// players are in, matching the no-active-session flow rather than starting the
-// game outright. Only the host may call it, and only when no game is in flight.
-// Returns the same errors as [Service.armRoomForHost].
-func (s *Service) ArmQuiz(ctx context.Context, joinCode string, hostPlayerID, quizID int64) error {
-	sess, err := s.armRoomForHost(ctx, joinCode, hostPlayerID, quizID)
-	if err != nil {
-		return err
-	}
-
-	// A newly armed quiz changes what the lobby shows (the Start controls appear),
-	// so signal subscribers to re-GET.
-	s.publish(sess.JoinCode, PhaseLobby)
-
-	s.logger.InfoContext(ctx, "live quiz armed",
-		slog.String(logJoinCodeKey, sess.JoinCode),
-		slog.Int64(logQuizKey, quizID),
-		slog.Int64(logHostKey, hostPlayerID))
-
-	return nil
-}
-
-// canArmQuiz reports whether a quiz can be armed to play in the room right now:
-// from an empty lobby that never started (the first game) or from the
-// between-games intermission (the next game). A game in flight (any other phase,
-// or a lobby already marked started) and a terminally finished room cannot.
-// Mirrors the RearmSession query's WHERE so the service rejects early with a
-// clear error, while the scoped UPDATE stays the real arbiter against a race.
-func canArmQuiz(sess *Session) bool {
-	if sess.Phase == PhaseIntermission {
-		return true
-	}
-
-	return sess.Phase == PhaseLobby && sess.StartedAt == nil
-}
-
-// EndSession closes a room for good (#836): the host control that terminally
-// finishes the room (finished + evict) from any live phase, so a host who is
-// done can shut it down rather than leaving it for the idle auto-close. Only the
-// host may call it. An already-finished room is treated as an idempotent no-op
-// (the host double-posted or a stale tab re-ended it). Returns
-// [ErrSessionNotFound] for an unknown code and [ErrNotHost] when the caller is
-// not the host.
-func (s *Service) EndSession(ctx context.Context, joinCode string, hostPlayerID int64) error {
-	sess, err := s.store.GetSessionByJoinCode(ctx, normalizeJoinCode(joinCode))
-	if err != nil {
-		return fmt.Errorf(errGetSessionByCodeFmt, err)
-	}
-	if sess.HostPlayerID != hostPlayerID {
-		s.logNonHostAttempt(ctx, "end", sess.JoinCode, hostPlayerID)
-
-		return ErrNotHost
-	}
-	if sess.Phase == PhaseFinished {
-		return nil
-	}
-
-	if err = s.store.Finish(ctx, sess.ID); err != nil {
-		return fmt.Errorf("failed to finish session: %w", err)
-	}
-
-	// The room is now terminal; signal subscribers to re-GET so every surface
-	// lands on the finished state and the live clients tear down.
-	s.publish(sess.JoinCode, PhaseFinished)
-
-	s.logger.InfoContext(ctx, "live session ended",
-		slog.String(logJoinCodeKey, sess.JoinCode),
-		slog.Int64(logHostKey, hostPlayerID))
-
-	return nil
-}
-
-// GetActiveSessionForHost returns the host's current active (non-finished) room,
-// or nil when the host has none (#836). Backs the "Resume hosting" link: a host
-// who opened a room up front and browsed away can return to it. Not gated beyond
-// the host id it is keyed on; the route layer host-gates the caller.
-//
-//nolint:nilnil // (nil, nil) is the deliberate "no active room" result; absence is not an error here.
-func (s *Service) GetActiveSessionForHost(ctx context.Context, hostPlayerID int64) (*Session, error) {
-	sess, err := s.store.GetActiveSessionForHost(ctx, hostPlayerID)
-	if err != nil {
-		return nil, fmt.Errorf(errGetActiveSessionFmt, err)
-	}
-
-	return sess, nil
-}
-
 // ArmStart arms the host's last-call countdown (the "Start in 60s" control):
 // it stamps the absolute deadline at now + the configured countdown, so every
 // surface renders the same server-clock countdown and the runner starts the
@@ -1237,46 +1020,6 @@ func (s *Service) Leave(ctx context.Context, joinCode string, playerID int64) er
 		slog.Int64(logPlayerKey, playerID))
 
 	return nil
-}
-
-// armRoomForHost validates that hostPlayerID hosts the room and that quizID is a
-// live quiz that can be armed now, then points the room at it via RearmSession
-// (which resets it to the lobby, not started) and returns the session. It does
-// not publish or start - the caller decides: [Service.ArmQuiz] stops here (the
-// room waits in the lobby), [Service.StartQuiz] marks it started. Returns
-// [ErrSessionNotFound] for an unknown code, [ErrNotHost] for a foreign host,
-// [quiz.ErrQuizNotFound] / [ErrNotLiveQuiz] for an unhostable quiz, and
-// [ErrGameInFlight] when a game is already running.
-func (s *Service) armRoomForHost(ctx context.Context, joinCode string, hostPlayerID, quizID int64) (*Session, error) {
-	sess, err := s.store.GetSessionByJoinCode(ctx, normalizeJoinCode(joinCode))
-	if err != nil {
-		return nil, fmt.Errorf(errGetSessionByCodeFmt, err)
-	}
-	if sess.HostPlayerID != hostPlayerID {
-		s.logNonHostAttempt(ctx, "arm", sess.JoinCode, hostPlayerID)
-
-		return nil, ErrNotHost
-	}
-	if !canArmQuiz(sess) {
-		return nil, ErrGameInFlight
-	}
-
-	qz, err := s.quizzes.GetQuiz(ctx, quizID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get quiz to arm: %w", err)
-	}
-	if qz.Mode != quiz.ModeLive {
-		return nil, ErrNotLiveQuiz
-	}
-
-	// RearmSession is scoped to "no game in flight", so it is the real arbiter if
-	// the room left that state between the read above and this write (a concurrent
-	// arm or a start that raced this one); it returns ErrGameInFlight to the loser.
-	if err = s.store.RearmSession(ctx, sess.ID, qz.ID); err != nil {
-		return nil, fmt.Errorf("failed to arm session: %w", err)
-	}
-
-	return sess, nil
 }
 
 // lobbyQuiz loads the room's quiz for the session state, or (nil, nil) for an empty
@@ -1548,29 +1291,6 @@ func (s *Service) allocateJoinCode(ctx context.Context) (string, error) {
 	}
 
 	return "", ErrJoinCodeUnavailable
-}
-
-// logNonHostAttempt logs an Info line for a non-host caller trying a
-// host-gated control, naming the action and who attempted it. A host who
-// cannot start or end their room is a thing the host wants to see explained,
-// not just a bare 403 in the access log.
-func (s *Service) logNonHostAttempt(ctx context.Context, action, joinCode string, playerID int64) {
-	s.logger.InfoContext(ctx, "live session control rejected: not host",
-		slog.String("action", action),
-		slog.String(logJoinCodeKey, joinCode),
-		slog.Int64(logPlayerKey, playerID))
-}
-
-// logAnswerNotOpen logs a Debug line for an answer rejected because no question
-// is open for it, naming the reason (wrong-phase / out-of-window /
-// option-mismatch). Debug because a late tap can repeat the line per stray
-// submit, while a host who wants to know why a pick did not land still has it.
-func (s *Service) logAnswerNotOpen(ctx context.Context, sess *Session, playerID int64, reason string) {
-	s.logger.DebugContext(ctx, "answer rejected: question not open",
-		slog.String(logJoinCodeKey, sess.JoinCode),
-		slog.Int64(logPlayerKey, playerID),
-		slog.String(logPhaseKey, string(sess.Phase)),
-		slog.String(logReasonKey, reason))
 }
 
 // publish fans out a session tick if a publisher is wired. The single
